Refactor base url to enable sort query parameter and add corresponding JSX
