+++ src/Components/portfolio/Portfolio.jsx
@@ -1,11 +1,12 @@
+/* eslint-disable jsx-a11y/label-has-associated-control */
 import { useState, useEffect } from 'react';
 import { IoCloseSharp } from 'react-icons/io5';
 import techIcons from '../../utils/techIcons';
 import './portfolio.css';
 // import data from '../../utils/data';
-import fetchData from '../../utils/fetchdata';
+import fetchData from './fetchdata';
 
-const url = 'http://localhost:3000/api/v1/projects';
+const BASE_URL = 'http://localhost:3000/api/v1/projects';
 
 const truncateText = (text, maxLength) => {
   if (text.length <= maxLength) return text;
@@ -17,20 +18,24 @@ const Portfolio = () => {
   const [projects, setProjects] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(false);
+  const [sortOrder, setSortOrder] = useState('-title');
 
   useEffect(() => {
     const getProjects = async () => {
       setLoading(true);
-      const data = await fetchData(url);
+
+      const data = await fetchData(`${BASE_URL}?sort=${sortOrder}`);
+
       if (data) {
         setProjects(data.data);
+        setError(false);
       } else {
         setError(true);
       }
       setLoading(false);
     };
     getProjects();
-  }, []);
+  }, [sortOrder]);
 
   if (error) {
     return <h2 className="error">Something went wrong...</h2>;
@@ -49,6 +54,22 @@ const Portfolio = () => {
     <section id="portfolio">
       <h5>My Recent Work</h5>
       <h2>Portfolio</h2>
+
+      <div className="container">
+        <div className="portfolio__controls">
+          <label htmlFor="sort">Sort by Title:</label>
+          <div className="select-wrapper">
+            <select
+              id="sort"
+              value={sortOrder}
+              onChange={(e) => setSortOrder(e.target.value)}
+            >
+              <option value="-title">Desc</option>
+              <option value="title">Asc</option>
+            </select>
+          </div>
+        </div>
+      </div>
 
       <div className="container portfolio__container">
         {projects.map((project) => (
