+++ src/Components/portfolio/portfolio.css
@@ -53,7 +53,34 @@
   margin: 2rem;
 }
 
-/***
+.portfolio__controls {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  margin-bottom: 2rem;
+  gap: 0.5rem;
+}
+
+.select-wrapper {
+  position: relative;
+}
+
+.select-wrapper select {
+  appearance: none;
+  padding: 0.5rem 2rem 0.5rem 0.8rem;
+  border-radius: 0.4rem;
+  border: 1px solid #ccc;
+  cursor: pointer;
+}
+
+.chevron-icon {
+  position: absolute;
+  right: 8px;
+  top: 50%;
+  transform: translateY(-50%);
+  pointer-events: none;
+} /***
+
 *! ===================== MEDIA QUERIES (MEDIUM DEVICES) ========================
 ***/
